Update the banner update
