+++ src/css/home.scss
@@ -23,13 +23,6 @@
     background-position: top -225px center;
     background-repeat: no-repeat;
 
-    .has-alert & {
-        margin-top: -130px;
-        padding-top: 100px;
-        z-index: -1;
-    }
-
-
     h1 {
         margin: 2.5rem 0 0.5em;
         font-size: 2.5rem;
